+++ spinta/formats/helpers.py
@@ -2,7 +2,7 @@ from typing import Iterator
 from typing import List
 from typing import Optional
 
-from spinta.backends import SelectTree
+from spinta.backends import SelectTree, get_property_base_model
 from spinta.backends import get_model_reserved_props
 from spinta.backends.helpers import get_ns_reserved_props
 from spinta.backends.helpers import get_select_prop_names
@@ -12,7 +12,9 @@ from spinta.components import Action
 from spinta.components import Context
 from spinta.components import Model
 from spinta.components import UrlParams
-from spinta.types.datatype import Array, ExternalRef
+from spinta.types.datatype import Array
+from spinta.types.datatype import Inherit
+from spinta.types.datatype import ExternalRef
 from spinta.types.datatype import DataType
 from spinta.types.datatype import Object
 from spinta.types.datatype import File
@@ -73,7 +75,22 @@ def _get_dtype_header(
             ):
                 name_ = name + '.' + prop.name
                 yield from _get_dtype_header(prop.dtype, sel, name_)
-
+    elif isinstance(dtype, Inherit):
+        if select and select != {'*': {}}:
+            properties = {}
+            for sel in select.keys():
+                base_model = get_property_base_model(dtype.prop.model, sel)
+                properties.update(base_model.properties)
+            for prop, sel in select_only_props(
+                dtype.prop,
+                properties.keys(),
+                properties,
+                select,
+            ):
+                name_ = name + '.' + prop.name
+                yield from _get_dtype_header(prop.dtype, sel, name_)
+        else:
+            yield name
     else:
         yield name